Rename file and update props.
